# [IOTDB-829] Accelerate delete timeseries