--- a/server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java
+++ b/server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java
@@ -452,8 +452,7 @@ public class StorageEngine implements IService {
     try {
       for (PartialPath storageGroupPath : IoTDB.metaManager.getStorageGroupPaths(path)) {
         StorageGroupProcessor storageGroupProcessor = getProcessor(storageGroupPath);
-        PartialPath newPath = path.alterPrefixPath(storageGroupPath);
-        storageGroupProcessor.delete(newPath, startTime, endTime);
+        storageGroupProcessor.delete(path, startTime, endTime);
       }
     } catch (IOException | MetadataException e) {
       throw new StorageEngineException(e.getMessage());
@@ -465,13 +464,10 @@ public class StorageEngine implements IService {
    */
   public void deleteTimeseries(PartialPath path)
       throws StorageEngineException {
+    StorageGroupProcessor storageGroupProcessor = getProcessor(path);
     try {
-      for (PartialPath storageGroupPath : IoTDB.metaManager.getStorageGroupPaths(path)) {
-        StorageGroupProcessor storageGroupProcessor = getProcessor(storageGroupPath);
-        PartialPath newPath = path.alterPrefixPath(storageGroupPath);
-        storageGroupProcessor.delete(newPath, Long.MIN_VALUE, Long.MAX_VALUE);
-      }
-    } catch (IOException | MetadataException e) {
+      storageGroupProcessor.delete(path, Long.MIN_VALUE, Long.MAX_VALUE);
+    } catch (IOException e) {
       throw new StorageEngineException(e.getMessage());
     }
   }