--- a/server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java
+++ b/server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java
@@ -449,12 +449,10 @@ public class StorageEngine implements IService {
 
   public void delete(PartialPath path, long startTime, long endTime)
           throws StorageEngineException {
+    StorageGroupProcessor storageGroupProcessor = getProcessor(path);
     try {
-      for (PartialPath storageGroupPath : IoTDB.metaManager.getStorageGroupPaths(path)) {
-        StorageGroupProcessor storageGroupProcessor = getProcessor(storageGroupPath);
-        storageGroupProcessor.delete(path, startTime, endTime);
-      }
-    } catch (IOException | MetadataException e) {
+      storageGroupProcessor.delete(path, startTime, endTime);
+    } catch (IOException e) {
       throw new StorageEngineException(e.getMessage());
     }
   }