--- a/server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java
+++ b/server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java
@@ -449,10 +449,14 @@ public class StorageEngine implements IService {
 
   public void delete(PartialPath path, long startTime, long endTime)
           throws StorageEngineException {
-    StorageGroupProcessor storageGroupProcessor = getProcessor(path);
     try {
-      storageGroupProcessor.delete(path, startTime, endTime);
-    } catch (IOException e) {
+      for (PartialPath storageGroupPath : IoTDB.metaManager.getStorageGroupPaths(path)) {
+        StorageGroupProcessor storageGroupProcessor = getProcessor(storageGroupPath);
+        PartialPath newPath = path.alterPrefixPath(storageGroupPath);
+        System.out.println("delete path :" + newPath + " from " + startTime + " to " + endTime);
+        storageGroupProcessor.delete(newPath, startTime, endTime);
+      }
+    } catch (IOException | MetadataException e) {
       throw new StorageEngineException(e.getMessage());
     }
   }
@@ -462,10 +466,13 @@ public class StorageEngine implements IService {
    */
   public void deleteTimeseries(PartialPath path)
       throws StorageEngineException {
-    StorageGroupProcessor storageGroupProcessor = getProcessor(path);
     try {
-      storageGroupProcessor.delete(path, Long.MIN_VALUE, Long.MAX_VALUE);
-    } catch (IOException e) {
+      for (PartialPath storageGroupPath : IoTDB.metaManager.getStorageGroupPaths(path)) {
+        StorageGroupProcessor storageGroupProcessor = getProcessor(storageGroupPath);
+        PartialPath newPath = path.alterPrefixPath(storageGroupPath);
+        storageGroupProcessor.delete(newPath, Long.MIN_VALUE, Long.MAX_VALUE);
+      }
+    } catch (IOException | MetadataException e) {
       throw new StorageEngineException(e.getMessage());
     }
   }