--- a/server/src/main/java/org/apache/iotdb/db/engine/memtable/AbstractMemTable.java
+++ b/server/src/main/java/org/apache/iotdb/db/engine/memtable/AbstractMemTable.java
@@ -21,15 +21,18 @@ package org.apache.iotdb.db.engine.memtable;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import org.apache.iotdb.db.conf.IoTDBConstant;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.engine.modification.Deletion;
 import org.apache.iotdb.db.engine.modification.Modification;
 import org.apache.iotdb.db.engine.querycontext.ReadOnlyMemChunk;
 import org.apache.iotdb.db.exception.WriteProcessException;
 import org.apache.iotdb.db.exception.query.QueryProcessException;
+import org.apache.iotdb.db.metadata.PartialPath;
 import org.apache.iotdb.db.qp.physical.crud.InsertRowPlan;
 import org.apache.iotdb.db.qp.physical.crud.InsertTabletPlan;
 import org.apache.iotdb.db.rescon.TVListAllocator;
@@ -230,20 +233,27 @@ public abstract class AbstractMemTable implements IMemTable {
   }
 
   @Override
-  public void delete(String deviceId, String measurementId, long startTimestamp, long endTimestamp) {
-    Map<String, IWritableMemChunk> deviceMap = memTableMap.get(deviceId);
-    if (deviceMap != null) {
-      IWritableMemChunk chunk = deviceMap.get(measurementId);
-      if (chunk == null) {
-        return;
+  public void delete(PartialPath originalPath, PartialPath devicePath, long startTimestamp, long endTimestamp) {
+    Map<String, IWritableMemChunk> deviceMap = memTableMap.get(devicePath.getFullPath());
+    if (deviceMap == null) {
+      return;
+    }
+    if (originalPath.getNodes().length <= devicePath.getNodes().length ||
+        originalPath.getTailNode().equals(IoTDBConstant.PATH_WILDCARD)) {
+      deviceMap.clear();
+    } else {
+      Iterator<Entry<String, IWritableMemChunk>> iter = deviceMap.entrySet().iterator();
+      while (iter.hasNext()) {
+        Entry<String, IWritableMemChunk> entry = iter.next();
+        IWritableMemChunk chunk = entry.getValue();
+        if (entry.getKey().equals(originalPath.getTailNode())) {
+          if (startTimestamp == Long.MIN_VALUE && endTimestamp == Long.MAX_VALUE) {
+            iter.remove();
+          }
+          int deletedPointsNumber = chunk.delete(startTimestamp, endTimestamp);
+          totalPointsNum -= deletedPointsNumber;
+        }
       }
-      // If startTimestamp == Long.MIN_VALUE && endTimestamp == Long.MAX_VALUE,
-      // it means that the whole timeseries is deleted
-      if (startTimestamp == Long.MIN_VALUE && endTimestamp == Long.MAX_VALUE) {
-        deviceMap.remove(measurementId);
-      }
-      int deletedPointsNumber = chunk.delete(startTimestamp, endTimestamp);
-      totalPointsNum -= deletedPointsNumber;
     }
   }