JIRA issue: https://issues.apache.org/jira/browse/IOTDB-829

### Local Performance Test result

#### Delete test result: 
- 1 device with 100k timeseries，original cost is 500s, after the improvement it reduce down to 100ms.
- 1 device with 300k timeseries. Before improvement it costs 1900s. While with the improvement, the cost reduce down to 210ms

#### Influence to querying:
100 timeseries，inserted 10000 points and executed M deletes.
We tested 1000000 queries with SQL:
``` select sensor_name from device_name```

- M = 5000
  Before delete improvement: average cost 19729ms
  After delete improvement: average cost 20901ms
  increased 5.9%
- M = 10000
  Before delete improvement:average cost 21185ms
  After delete improvement: average cost 22414ms
  Increased 5.8%

### Re-tested after revision in 10.15
Test SQL
```
delete timeseries root.sg1.d1.*
```
#### Delete test result: 
- 100k timeseries under root.sg1.d1. master branch cost is 632s, the cost of this PR reduce down to 76ms.
- 300k timeseries under root.sg1.d1. master branch cost is 2040s. the cost of this PR reduce down to 111ms